requirements.txt
beautifulsoup4
Mastodon.py
redis
requests
